Clean up some unused variables
