+++ src/GamePos.ts
@@ -1,0 +1,13 @@
+
+export class GamePos {
+    public id: number;
+    public x: number;
+    public y: number;
+  
+    constructor(id: number, x: number, y: number) {
+      this.id = id;
+      this.x = x;
+      this.y = y;
+    }
+}
+
